Refactor presentation
