+++ app/build.gradle
@@ -39,6 +39,7 @@ dependencies {
     }
     implementation "android.arch.lifecycle:extensions:$lifecycle_version"
     implementation 'com.beust:klaxon:3.0.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:0.22.5'
 }
 repositories {
     mavenCentral()
