+++ app/src/main/res/layout/activity_main.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <fragment
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:name="com.theleftbit.gifwallet.presentation.GifListFragment"
+    android:name="com.theleftbit.gifwallet.presentation.giflist.GifListFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"/>
